runtime fix 7

+++ snake_game/src/snake.cpp
@@ -33,7 +33,7 @@ int Snake::start_game() {
     return (1);
     if (!(fpga_v_addr = map_physical (snake_fd, LW_BRIDGE_BASE, LW_BRIDGE_SPAN)))
     return (1);
-    snake_v_addr = (int *) ( (int)fpga_v_addr + SNAKE_GAME_BASE);
+    snake_v_addr = (int *) ( fpga_v_addr + SNAKE_GAME_BASE);
     LEDR_ptr = (int *) (fpga_v_addr + LEDR_BASE);
     cout << "FPGA virtual addr is " << hex << (int) fpga_v_addr << endl;
     cout << "Snake virtual addr is " << hex << (int) snake_v_addr << endl;
